+++ ngClient/App/controllers/account/loginController.js
@@ -3,9 +3,9 @@
 
     var controllerId = 'LoginController';
     angular.module('main')
-        .controller(controllerId, ['userFactory', '$location', '$rootScope', 'serviceAppUrl', 'logger', LoginController]);
+        .controller(controllerId, ['userFactory', '$location', 'locationHistory', 'serviceAppUrl', 'logger', LoginController]);
 
-    function LoginController(userFactory, $location, $rootScope, serviceAppUrl, logger) {
+    function LoginController(userFactory, $location, locationHistory, serviceAppUrl, logger) {
         logger = logger.forSource(controllerId);
 
         var vm = this;
@@ -28,9 +28,8 @@
         function getAccessToken() {
             userFactory.getAccessToken(vm.email, vm.password)
                 .success(function () {
-
-                    // Redirect the user to the previous page
-                    $location.path($rootScope.locationHistory[$rootScope.locationHistory.length - 2].path());
+                    var locationItem = locationHistory.get(locationHistory.getHistory().length - 2);
+                    $location.url(locationItem.url());
                 })
                 .error(function (response) {
                     if (typeof response.error_description !== 'undefined') {